Merged PR 5589: Fix: Promo code (and the whole cart ) is kept in spoofed user
There is no connection between the user and the cart (also not a logged-in user can create a cart). Promo codes (campaigns) are included in the cart which is preserved in a cookie. One of the solutions is to remove the cart whenever we try to log in as a spoofed user.

- reset cart when logging in as a spoofed user in the pages account and preview
- remove code duplication (use a declared method instead)

Related work items: #20612